Split up lf_helpers in docs.
Also cleans up the functions in __init__.py into a utils.py file.
