+++ docs/user/lf_helpers.rst
@@ -3,11 +3,39 @@ Labeling Function Helpers
 
 This page shows descriptions of the helper functions included with Fonduer_
 which can be used to label candidates based on textual, structural, tabular,
-and visual information.
+and visual information. We group each lf_helper based on the modality of
+information that they leverage.
 
 ----
 
-.. automodule:: fonduer.supervision.lf_helpers
+General Labeling Function Helpers 
+---------------------------------
+
+.. automodule:: fonduer.supervision.lf_helpers.utils
+    :members:
+
+Textual Labeling Function Helpers
+---------------------------------
+
+.. automodule:: fonduer.supervision.lf_helpers.textual
+    :members:
+
+Structural Labeling Function Helpers
+---------------------------------
+
+.. automodule:: fonduer.supervision.lf_helpers.structural
+    :members:
+
+Tabular Labeling Function Helpers
+---------------------------------
+
+.. automodule:: fonduer.supervision.lf_helpers.tabular
+    :members:
+
+Visual Labeling Function Helpers
+---------------------------------
+
+.. automodule:: fonduer.supervision.lf_helpers.visual
     :members:
 
 .. _Fonduer: https://github.com/HazyResearch/fonduer
